fixed azure storage, reorganized validation func
